refactor: fold styles/ and presentation/ into components/ui/ and domain/
Icons.tsx moves to components/ui/ (the first entry there, ahead of
the broader components/ reorg); tier-styling.ts moves to
domain/company/services/, alongside company-tier-access.ts, since
it's Company-tier-scoped presentation logic rather than a standalone
concern.
