+++ src/components/Companies/CompaniesContainer/index.tsx
@@ -5,7 +5,10 @@ import { motion } from "framer-motion";
 
 import Company, { CompanyProps } from "@/components/Companies/Company";
 import { CompanyTier } from "@/domain/company/company-tier";
-import { getTierStyling, tierDisplayLabel } from "@/presentation/tier-styling";
+import {
+  getTierStyling,
+  tierDisplayLabel,
+} from "@/domain/company/services/tier-styling";
 
 interface CompaniesContainerProps {
   tier: CompanyTier;
