+++ src/components/Companies/CompanySocials/index.tsx
@@ -2,8 +2,6 @@
 
 import React from "react";
 
-import HeadingText from "@/components/HeadingText";
-import SocialMediaCard from "@/components/SocialMediaCard";
 import {
   Facebook,
   Globe,
@@ -11,7 +9,9 @@ import {
   Linkedin,
   Twitter,
   Youtube,
-} from "@/styles/Icons";
+} from "@/components/ui/Icons";
+import HeadingText from "@/components/HeadingText";
+import SocialMediaCard from "@/components/SocialMediaCard";
 
 interface CompanySocialsProps {
   twitterLink: string | undefined;
