+++ src/domain/company/services/tier-styling.ts
